0.7.2 release: * Create `openmaps_auth.utils` and `openmaps_auth.tls` modules and reorganize some functions and classes. * Add style and icons to support the use of Django's messages. * Upgrade Django to 4.1.3.
