+++ openmaps_auth/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.7.1"
+__version__ = "0.7.2"
