+++ requirements/base.in
@@ -1,5 +1,5 @@
 beautifulsoup4==4.11.1
-Django==4.1.2
+Django==4.1.3
 django-environ==0.9.0
 gunicorn==20.1.0
 gevent==22.10.1
